新增EmployeeController、EmployeeDAO、EmployeeService、EmployeeServiceTest

+++ showcase/showcase-biz/src/main/java/com/justdoit/showcase/airport/entity/Employee.java
@@ -7,7 +7,7 @@ import javax.persistence.Entity;
 import com.justdoit.showcase.base.entity.BaseEntity;
 
 @Entity
-public class Employee extends BaseEntity<Integer> {
+public class Employee extends BaseEntity<Long> {
 
 	/**
 	 *
